requirements.txt
amqp==2.6.1
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
black==20.8b1
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.7
Django==3.1.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.1
drf-yasg==1.17.1
flake8==3.8.3
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==20.4
pathspec==0.8.0
prompt-toolkit==3.0.7
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
serializers==0.2.4
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
zipp==3.2.0
